ファイルの削除と整理 / API用環境変数の設置
