+++ next.config.js
@@ -1,3 +1,6 @@
 module.exports = {
   reactStrictMode: true,
+  env: {
+    NEXT_PUBLIC_RESAS_API_KEY: process.env.NEXT_PUBLIC_RESAS_API_KEY
+  }
 }
