+++ src/pages/index.tsx
@@ -1,0 +1,26 @@
+import Head from 'next/head'
+import styles from '../styles/Home.module.css'
+
+export default function Home() {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <header className={styles.header}>
+        <h1>人口推移</h1>
+      </header>
+
+      <main className={styles.main}>
+        <p>Hello</p>
+      </main>
+
+      <footer className={styles.footer}>
+        <small>Powered by Vercel / Next.js / RESAS API / Highchart</small>
+      </footer>
+    </div>
+  )
+}
