+++ src/styles/Home.module.css
@@ -1,0 +1,11 @@
+.header{
+  background: #ccc;
+}
+
+.main {
+  background: #ddd;
+}
+
+.footer {
+  background: #eee;
+}
