+++ src/styles/globals.css
@@ -13,4 +13,6 @@ a {
 
 * {
   box-sizing: border-box;
+  margin: 0;
+  padding: 0;
 }
